test: consolidate _init_tile_pool_for_tests into conftest.py
Five files carried byte-identical asyncpg pool setup for tile-serving
tests (DSN derivation, pool sizing, the too-many-clients retry wrap,
and pool_module._tile_pool mutation). A retry fix previously had to be
hand-applied at three of the five copies separately.

Move the fixture into tests/conftest.py as a single non-autouse
definition; the four class-scoped consumers keep their existing
@pytest.mark.usefixtures. test_embed_tokens.py's file-scoped
autouse=True becomes a module-level pytestmark with the same mark,
preserving its request-based guard that skips pool creation for tests
touching no DB fixture.
