+++ backend/tests/test_tile_gzip_offload_perf005.py
@@ -26,7 +26,6 @@ import gzip
 import uuid
 from unittest.mock import patch
 
-import asyncpg
 import pytest
 from httpx import AsyncClient
 from sqlalchemy import text
@@ -34,7 +33,6 @@ from sqlalchemy import text
 from app.core.config import settings
 from app.modules.catalog.datasets.domain.models import Dataset, Record
 
-from tests.conftest import _run_with_too_many_clients_retry
 from tests.factories import get_user_id
 
 
@@ -102,21 +100,6 @@ async def _create_point_data_table(session, table_name: str) -> None:
 async def _cleanup_data_table(session, table_name: str) -> None:
     await session.execute(text(f"DROP TABLE IF EXISTS data.{table_name}"))
     await session.commit()
-
-
-@pytest.fixture
-async def _init_tile_pool_for_tests():
-    """Create a real asyncpg pool for tile rendering (lifespan does not run under ASGITransport)."""
-    import app.processing.tiles.pool as pool_module
-
-    dsn = settings.test_database_url.replace("postgresql+asyncpg://", "postgresql://")
-    pool = await _run_with_too_many_clients_retry(
-        lambda: asyncpg.create_pool(dsn=dsn, min_size=1, max_size=3, command_timeout=10)
-    )
-    pool_module._tile_pool = pool
-    yield
-    await pool.close()
-    pool_module._tile_pool = None
 
 
 def _make_to_thread_spy():
