+++ backend/tests/test_tile_signing.py
@@ -13,7 +13,6 @@ import time
 import uuid
 from unittest.mock import patch
 
-import asyncpg
 import pytest
 from httpx import AsyncClient
 from sqlalchemy import text
@@ -22,7 +21,6 @@ from app.core.config import settings
 from app.modules.catalog.datasets.domain.models import Dataset, Record
 from app.processing.tiles.signing import generate_tile_signature
 
-from tests.conftest import _run_with_too_many_clients_retry
 from tests.factories import get_user_id
 
 
@@ -96,21 +94,6 @@ async def _create_data_table(session, table_name: str) -> None:
 async def _cleanup_data_table(session, table_name: str) -> None:
     await session.execute(text(f"DROP TABLE IF EXISTS data.{table_name}"))
     await session.commit()
-
-
-@pytest.fixture
-async def _init_tile_pool_for_tests():
-    """Initialize asyncpg pool for tile tests."""
-    import app.processing.tiles.pool as pool_module
-
-    dsn = settings.test_database_url.replace("postgresql+asyncpg://", "postgresql://")
-    pool = await _run_with_too_many_clients_retry(
-        lambda: asyncpg.create_pool(dsn=dsn, min_size=1, max_size=3, command_timeout=10)
-    )
-    pool_module._tile_pool = pool
-    yield
-    await pool.close()
-    pool_module._tile_pool = None
 
 
 # ---------------------------------------------------------------------------
